Get install more or less working.
We still have work to do in terms of making this robust but the basic features are:
* Start entire install process using NPM in the www/jxcore directory of the Cordova app.
* The install will default to downloading the JS files from NPM and everything else from Thali_CordovaPlugin's depot.
* There is an override we can use during development to let us download the JS and Cordova contents into the app from a local directory (see the TeamReadMe.md)
* We now download jxcore directly from GitHub and target the right branch
* If there is an update in NPM then we can pick it up with jx npm update and our code will do the right thing in terms of checking what else has changed and automatically updating the Cordova app.
* We have an after prepare script that handles some clean up in terms of copying files and a few edits. This script will also delete the special install directory we create with our install logic so it doesn't get installed onto the mobile device
* We will now install the btconnectorlib2 first from local maven and then from remote so users don't need to install maven anymore but BT developers can still work with the project.

I need to run more tests to make sure this thing is robust. In particular I'm running into a bizarre error stack during build with Cordova that doesn't seem to break anything but it seems to mean that the build will sometimes randomly fail for no obvious reason. I'm investigating. I also need to do a full pass on all the code and do lots of clean up.