Proposals for changes, please read
.gitignore - We shouldn't check in test/www/jxcore/thali since
that just repeats files that are already in the project.

build/deploytest.js - I moved this into the test directory since
it is about testing and it has paths that depend on where its
launched.

package.json - I removed this from the root of the project because
it confuses the heck out of NPM and we don't need it there since
none of our code is in the root.

mockmobile - I moved it to the test directory since it doesn't need
to be in production.

thaliscenarios.js - I don't have time now to finish this so I would
ask you to. But I introduced a structure function called
connectWithRetryTestAndDisconnect that can be used as a scafold
for all the remaining tests in the file so we don't keep on repeating
exactly the same code over and over.

thalireplicationmanager - It didn't properly handle double starts
and it wasn't cleaning itself up on stop failures.


README.md -