Fixed test to work as advertised
The new version of the first test in thaliscenarios assumes that we should
get a networkChanged event when stopBroadcasting is called but that isn't
how the API is supposed to work. The code also depended on a race condition
where the call to startBroadcasting/stopBroadcasting and all their callbacks
where in a tight loop that would effectively allow them to run serially
even though they really should be in parallel.