identityExchange.md - Updated to exhaustively define exactly what the protocol needs to do and how it needs to do it. I'm not quite done yet but this is a lot of work so I want to back it up.
thalinreplcatiomanager.md - Removing peerIdentityExchange which I moved
to identityExchange.md (where it always should have been) and aded
connectionSuccess which is a temporary hack we need for identity
exchange.

jxcore/app.js - Hosigin the thaliemitter require and declaring a variable
that wasn't properly declared.

identityExchange.js - Just putting the body parser in on specific paths
so that nobody has to remember to add this to their express server. We'll
handle it ourselves.

package.json - Adding body-parser as a dependency

thalireplicationmanager.js - Moving addListener to before startBroadcasting
so we don't miss any events. And a little bit of code cleanup.