Yet another backup
Added the ability to start and stop thaliReplicationManager using promises
because the event based code was just too complicated. Updated both
SmallerHashStateMachine and the first part of identityexchange to use it.

Fixed a bug where I called the start method for replication manager without
any of the required arguments! Added those values to the constructor
for SmallerHashStateMachine.

Note that there is a known error in here due to asynchronous code
introduced to SmallerHashStateMachine. I will fix it tomorrow.