Another backup
All tests (which are mocked based) except for a few based on nock which
doesn't work on JXcore (I've filed a bug) are now passing.

Fixed a race condition between receiving a peer exchange available event
and getting the start callback. Added a test for that.

Wrote a "live" test that hasn't passed yet but so far I think it's because
service discovery isn't working on my phones, need to reboot.

Had to add my own buffer compare function since node 0.10 doesn't have it.
Added tests for that too.

Fixed generateSliceSHA256Hash which we use to generate our local device
ID to properly produce a sliced buffer. We were slicing the base64 value!
This produced meaningless content.