Enabling mocks to run live
In theory (we haven't tested it yet, I'm checking in so Matt can test
with me) this enables all the testIdentityExchangeEndpoint tests to
run "live" when on mobile and "mock" when on desktop. Also fixed a minor
bug that the peerIdentityExchange event comes from the identityExchange
object not the replication manager.