Almost there!
Every single test is passing with the exception of one test that is
failing two asserts and I know why. I'll fix it tomorrow.

Found a endless loop in the Thali Replication Manager, once it starts
retrying a peer it won't ever stop, even if the replication manager is
turned off!

A lot of the changes here are because Matt put in an editor file that
enforces comment editor settings so anytime I open a file that doesn't
match those settings it gets automatically reformated.

Desktop setup should now be working