Making a disaster of tear down
I was running into all sorts of nasty issues with tests not running right
so I decided to re-write the teardown code in testidentityExchangeEndpoint
and that was good and bad. It was good in that all the tests now pass.
It was bad in that tests now take an eternity to run and at the end
of testIdentityExchangeEndpoint the final test literally never returns
from server.close so no other tests run. I tried to close all the HTTP
client connections which I assume are holding the server instance open
but that doesn't seem to really help.