Don't exit the Cordova process in a hook.
The Android hook had an async step after which the process was exited. This
caused the Cordova process to exit too soon and was seen, for example, when
`cordova run android` was tried to be executed.

The `appRoot` variable is taken from the Cordova context instead of from
`__dirname` so that the path gets right even if the plugin is added via a
symbolic link.

Also removed unneeded dependency to the `lie` module.

Fixes #143.