Exit the setup scripts if any sub-command fails.
The biggest reason for doing this is that these scripts are called during the
CI build process and if the scripts don't overall exit with a non-zero exit
code, the CI thinks the build worked. The result was that even when parts of
the setup failed, the CI thought the build was successful.

Cordova plugin installation output is now logged for easier debugging.