Use OS temporary folder as the download cache.
A part of the installation is the download of the required version of the
JXCore Cordova plugin (currently ~90MB download) and unpacking that with `jx`
before adding it as Cordova plugin using Cordova commands.

This commit implements caching of the unpacked plugin folder so that the
overall installation process becomes faster. The OS temporary folder is used
for caching instead of any Thali-app folder so that full benefits of the cache
are received even if the app-folder is re-generated (this is the case with CI
scripts, for example).

This commit also adds some catches to the promise-chains so that in case of
errors, it is easier to see where the issue was.

Relates to #192.