Simplification related to previous installations.
The installation script used to store the previously installed plugin version
to a separate file so that a version comparison was possible between what
was installed and what was required.

This is seen as an unnecessary complication to the installation process so
this commit removes this functionality and always installs the version of the
JXCore Cordova plugin that this version of Thali requires.