Leverage a remote cache in CI.
The CI environment builds on a virtual machine that boots from a clean state
and thus all local caches for downloads are empty when the build starts.

In this commit, the caching logic is extended to support a scenario where in
the CI, we use a remote cache over scp to prevent big external downloads from
slowing down the overall build process.

Fixes #192.