Updated thaliMobile spec based on feedback
package.json - Changed the jsdoc command to a public and internal version
in order to hide private methods like killConnection.

thaliMobile.js - Emptied this out, it will become the new API that we
expect to call directly rather than the nativeCall et al. which we
will wrap.

thaliMobileNative.js - This is the new formal definition for nativeCall
and friends.