A test and a fix for sending data.
Implemented a test for the data sending test. The test revealed an issue with
how the acknowledgment was done in the server that receives the data and
this commit also fixes that.

The issue was caused by the fact that data may be sometimes sent over the
socket connection in "batches" so the client may actually receive multiple
acknowledgments at the same data callback.

This commit also adds timestamp-based logging to the client and server, which
makes it easier to debug issues in CI environment.