Working on notification client
thaliMobile - Updated it to specify that it has primary responsibility for
giving a clean indication of when peers are and are not responsible. In
other words it has to apply whatever heuristics we think we need to guess
when someone is and is not there. It also has to keep the
peerAvailabilityChanged feed free of duplicates and such.

thaliNotificationClient - Beefed up the text explaining how this is
supposed to work. Still a bunch of work to do.

thaliPeerPoolInterface - Added a dequeue method and clarified a few things.