.jscsrc - We need to allow for unlimited errors so that IDEs like intellij can show all errors.
thaliMobile - Clarify how we handle various events in relation to
declaring when peers are not around. Also clarifying some heuristics on
how to handle declaring peers not present when connections are failing.

thaliMobileNativeWrapper - Just a little clean up to get it closer to
passing JSCS and JSHINT

thaliNotificationAction - Refactoring and cleanup

thaliNotificationClient - Refactoring and bug fixes

thaliPeerAction - Factored this into its own file

thaliPeerDictionary - Factored this into its own file

thaliPeerPoolInterface - Refactoring

package.json - Just cleaning up key words

thaliNotificationBeacons - Changed this from being a class to being a
series of static methods that we expose directly. Note that to get jsdoc
to be happy we have to define the functions directly on module.exports
or jsdoc doesn't see them.