Implement mock for the Connect call.
This is enough to allow the Thali replication manager test to pass on desktop.

When running the test, the cleanup doesn't currently happen properly and even
after a successful run, open sockets are left behind. These can be seen with
tools like `lsof` and prevents the test process from exiting.