Fix an issue that prevented running two tests.
The timeouts in the server were set in a way that a timeout from a previous
test got triggered when the next test was tried to be run and thus cut off
the execution of the second test.

This commit also fixed some log prints and the find peers test in case it
uses Bluetooth address lists.