Changed Android hook to be run before compile.
The previous configuration was to run the hook after prepare. This approach
seemed to not work in cases where plugins are added before platforms, because
then, the after prepare hook got triggered before all the plugins were
installed. This meant that the file replacement done in the hook did not work
because the source and target files might not have existed yet.

The before compile hook will be always run after all plugins and platforms
are added so there file replacement should always work.

Fixes #431.