Removed a duplicate readme.
This had been committed to git repository by accident.

The idea is that the readme in the root of this repository is copied over
in the npm prepublish script. That way, the same readme is visible in the
front of the github project and in npmjs to which the npm package is published.