Changes that @tobybrad and I agreed to
Specified that if connect is called a second time on the same peer ID and the first call either has resulted in a connection or hasn't completed yet then we will return an error. That way we don't have to manage a queue for connection calls at the native layer.

Fixed the link to the binding spec.

Specified that when we get too many peer Availability Changed updates that overloads our storage queue between sending batch updates to
node.js that we should drop the oldest announcements first in order to keep the queue at or below it's maximum size. Also clarified that for
incomingConnectionToPortNumberFailed only one notification in regard to any particular port's failure needs to be sent in each 100 ms window.

Specified that when the first subscription is received to networkChanged that we have to fire a networkChanged event with the current status. That
will let the thaliMobileNativeWrapper initialize it's method that tracks this information.