Enable the right hooks for peer pool interface
TCPServersManager - Put in a new event incomingConnectionState to track
incoming connections. This is needed mostly for Bluetooth. Also added a
method to allow for terminating incoming connections at the native layer
to control how many are active. Again, for Bluetooth.

thaliMobile - Added the getNetworkStatus method so clients can know if
they are online or not.

thaliMobileNative - Added cellular as an option for radioState since at
least in Android we can tell if we are on cellular or wifi.

thaliMobileNativeWrapper - Clarified that if the incoming TCP connection
from the TCP/IP layer is termninate than the underlying Bluetooth or
MPCF connection MUST be terminated as well. Also expose the ability to
terminate a native incoming connection here. As well as exposing the
incomingConnectionState event from TCPServersManager

thaliPeerAction - Clarifying kill semantics

thaliPeerPoolDefault - An outline of our default peer pool, it's not
terribly smart.

thaliPeerPoolInterface - Just clarifying that we have to hijack the kill
method on actions we are given.

thaliWiFiInfrastructure - clarifying how network status works on the
desktop with WiFi.