Tons of bug fixes
thaliNotificationClient - Got rid of addressBookCallback which was based
on an older design before we switched to tracking sequence numbers from
clients. Now we submit our list of peers we want to do pull requests to
start and call start again in order to update it.

thaliNotificationServer - Added an explicit start method to register the
beacon server router. Changed the public key type to buffer. Got rid of
the stop method since there isn't anyway to remove a route from a router
in Express.

thaliPullReplicationFromNotification - This class now creates its own
instance of ThaliNotificationClient instead of having it passed in. Changed
public key type to buffer. Changed setNotifications to be a start method
that takes as an argument all the peers we would want to try a pull
replication with. Introduced a stop method that stops listening for new
peer discoveries and shuts down all replication.

thaliReplicationManager - A new class that manages the all up replication
experience for users.

thaliReplicationPeerAction - Changed public key type to buffer.

thaliSendNotificationBasedOnReplication - It now creates its own
ThaliNotificationServer object. It now has a start method where we pass in
the peers we want to notify of updates.