Fixes for issues in code review
thaliMobileNativeWrapper - Specifying that we have to fire a
networkChangedNonTCP event when we get the first subscription so that we
can set state for the getNetworkStatus in thaliMobile.

thaliReplicationManager - Make thailPeerPoolInterface optional (we are so
going to regret that) and changed start to take one list of users and
clarified that the list is optional (although how useful a p2p system that
won't talk to anyone is, is another issue).

thaliWifiInfrastructure - Fixed two typos