Changed jx install back to jx npm install
We had originally moved to jx install because it's clear that this is what
JXCore tests with but it causes an issue because there is a conflict with
a file we have called install.js and in general it's good to be clear that
we are using npm.

I also added a note that folks should be using the V8 version of JXcore
on the desktop due to a known bug.