WIP
testPromiseQueue - Added tests for new enqueueAtTop functionality as well
as to make sure that promises that return values directly will work.

testThaliSendNotificationBasedOnReplication - Leaving myself a note on
a test to write.

promiseQueue - Enabled it to enqueue either at the top or bottom of the
queue.

thaliNotificationServer - Moving the comment about how the constructor
works to a section where it should show up in the constructor's doc

thaliSendNotificationBasedOnReplication - Further filling out the
functionality, still a WIP.