Fix issue with duplicate test names
Before, test names had to be unique across all test files or otherwise some
tests would be silently skipped.

The solution in this commit is to prepend a running number to the test name
so that it becomes unique.

Since all peers require exactly the same files in same order, the running
number should match between peers.

In `UnitTestFramework.js` only a non-functional change to clean up logging.