WIP
testThaliSendNotificationBasedOnReplication - Specified mock expectations
in a way that is closer to correct (but still wrong due to bugs in sinon)
and got rid of the fake clock because I think that Promises aren't working
with it but I'm not sure since I think the promise library we use does
use setTimeout, but I'm not sure.

thaliSendNotificationBasedOnReplication - Fixed a whole passel of race
conditions and there is at least one more that I know of. I filed that one
in the issue for tomorrow.