Do not require start for getNonTCPNetworkStatus
This was done mainly because getNonTCPNetworkStatus is used from
thaliWifiInfrastructure and if start would have been required, we
would have had to call that from thaliWifiInfrastructure, which
is not what we want. That is, because if someone wants to command
thaliWifiInfrastructure to start, it doesn't mean that automatically
thaliMobileNativeWrapper should be started as well.