Improved the tests in desktop environment
The Mobile mock is now capable to run most of the tests in
testThaliMobileNative.js. The main (and most challenging) part
is still missing and that is the mock for the connect call.

Refactored Wifi and Bluetooth toggling API in test utils to be promise-based
for a more convenient usage.