Crypto fixes
README - Just pointing out to avoid sudo if possible

testThaliNotificationBeacons - Changes notificationBeacons and long to
lower case variables since they are not constructors. Changed the constants
to use the values exposed by ThaliNotificationBeacons. Cleaned up and
expanded bad arguments test. Fixed timing tests so they now realize that
we pass in time until expiration, not absolute time, when creating
beacons. Add time test to take into account fuzzing

thaliNotificationBeacons - Made constants into exported values so we can
use them for testing. Changed one_day to show how it was derived. Adding in
fuzzing of expiration date. Changed seconds to milliseconds in
generatePremableAndBecaons and fixed the code so now the value is treated
as the expiration interval and not an absolute date. Removed IV because
crypto review showed it wasn't useful.