Avoid writes to stderr during test run
This commit was done so that in a success test run, we would not write to
stderr, because in some environments, such output is considered an error
behavior and fails the test.

The PouchDB.defaults call was moved so that it gets called only once, because
otherwise, event emitters gets subscribed to many time which caused a warning
like `EventEmitter memory leak detected` that was written to stderr.