Get rid of the radio toggling at the end of tests
The radio toggling at the beginning of the tests was removed as a workaround
for #563. That would be good to put back once a root cause and fix is found,
but even then, it should not be necessary to turn off radios at the end of
tests.

This was originally done for CI where devices didn't reboot in between tests,
but now devices do reboot and thus tests always start from clean radio states.