Reduce the severity in case of state mismatch
The mismatch can also happen in very normal scenario where stop is called
for thaliMobile, which in turn ends to calling both stops in
thaliWifiInfrastructure. After both stops are completed, there is an
independent emit of the discover state change and at the time of the first
one of those, the state doesn't yet match with the target.