Get tests passing again
testThaliTCPServersManager - linting

createNativeListener - Didn't realize that delete on an array just unsets
the value but not the key. So we have to use splice.

tcpServersManager - Removed DEFAULT_CONNECTION_TIMEOUT_IN_MS and instead
used a value from thaliConfig and put constants onto the prototype so
I didn't have to require tcpServersManager in creativeNativeListener which
created a nasty cyclical require situation. Also removed routerPort from
the call to createNativeListener because it really wasn't needed.