Make code compliant with connect
For some odd reason connect returns a string of a parsed JSON object
rather than an actual JSON object (which we can easily pass from native
to Javascript). Ville caught this and I had to update the code to both
produce in the tests and consume in the production code the right thing.

createNativeListener - Fixed a bug Ville pointed out in how we return errors

thaliMobileNative - Making a bizarre screw up into the official spec.

Linting - mockmobile.js, MockMobile