Take hardware capabilities into account in tests
If a device does not have required capabilities, the unit test app will
not continue and the device won't present itself to the coordination server.
Instead, it logs a success message to inform the CI that this completed
what it is supposed to do, i.e., load the app, query the capabilities and quit.