WIP
testThaliMobileNativeWrapper -
* changed ThaliMobileNativeWrapper to thaliMobileNativeWrapper since it
is not a constructor.
* Added catches in a few places to make sure that we will cleanly catch
the error and tear down the test so we don't screw up the whole test run.
* Added some empty tests that are set to fail in the forelorn hope that
they will be added
* Changed the incomingConnectionToPortNumberFailed test to work with the
event definition.

wifiBasedNativeMock - Lint

thaliMobileNativeWrapper -
* renamed global variables to start with 'g' so it would be easy to pick
them out when reading the code
* added an _isStarted function for testing
* lots of linting
* created handler to deal with failed connection events
* created handler to deal with failedrouterPort events
* centralized the code for starting and stopping servers manager
* change stop() method so that it will try to stop everything even if
some things fail and will always put itself into not started state
* moved the guts of stop() into its own method so we could re-use it
with certain errors
* added stub for terminateListener, this will be added later with
tcpServersManager
* cleaned up definition of incomingConnectionToPortNumberFailed to cover
both of the error cases and the possibility of multiple errors, also
defined that we stop the system when this happens