Enabling psk
thaliNotificationBeacons -
* changes parseBeacons to return a structure with
the raw values needed to generate the artifacts required to set up a PSK
connection.
* Changed buffer to Buffer to intellij type detection would
match properly.
* generatePskIdentity and generatePskSecret are used to generate the
named values, we'll need this for the notification client.

testThaliNotificationBeacons - Added tests for the two new functions,
updated an existing test for the new response to parseBeacons and got
rid of one test that was completely redundant