WIP - I got a basic scenario working
But then I broke it because I took a short cut where I didn't let the
wifi mock wrap the services manager listener. I need to fix that bug.

testThaliMobileNativeWrapper - Created a test that does a loop back HTTP
request via the wifi mock but doesn't require the coordinator. This is
wicked useful for debugging. Right now I broke it to try and wrap the
tcp services manager in the wifi mock which isn't working at the moment.

wifiBasedNativeMock - Cleaned up code that should be using pipes (which
automatically handle failure), catching errors and putting in paranoid
code to clean things up in case of hard closes with out ends.
   Added a method to let me emit my own location in a way that would be
visible to the rest of the system, I need this for when I want to do
loop back tests without the coordinator so I can easily debug.

createPeerListener - Fixed a typo that meant the code shouldn't have worked.