Get forward connections working!
thaliMobileNativeWrapper - We were pointing at the wrong port for
startUpdateAdvertisingAndListening so we skipped the mux layer, fixed.

testThaliMobileNativeWrapper - Had two variants of the same test, one that
runs on a local loop back so it's super easy to debug and another that
uses the coordinator. So I put the common code into a shared function and
wrapped the debug test in a mobile check since it's only intended for the
desktop.