Fix tests in testThaliMobileNativeWrapper
When we run the full coordinated test run, we end up running first tests in
testThaliMobileNative and afterwards the ones in testThaliMobileNativeWrapper.

The issue was that in testThaliMobileNative we call
Mobile('peerAvailabilityChanged').registerToNative during test runtime, which
ends up overwriting the handlers in testThaliMobileNativeWrapper.

This commit allows us to put back the right handlers when needed.