WIP
testCreatePeerListener - Added tests and fixed a bunch of minor issues.

testMakeIntoCloseAllServer - tested eatNotRunning

makeIntoCloseAllServer - Added 'eatNotRunning' option to eat the exception
we get if we close a server that hasn't started or is already closed.
This is useful when we have asynchronous or test code that needs to close
a server that might already be closed.

createNativeListener - Changed the structure of _pendingReverseConnections
so it uses properties instead of array positions

createPeerListener - Pretty major surgery including getting rid of
duplicated code, pulling functions out of their multiple levels of
nesting (which made it hard to read the code or tell where we are) and
fixing a variety of edge case errors.