Reducing wait time for reverse connections
We were waiting 10 seconds before which is a bit nuts since we only get
notified of the port to wait for if the system already has a connection
via that port. So the wait is just to deal with some minor race conditions
and so should be able to be much shorter.