#696 Fix Availability of Failed peers
If a peer connection fails we kill its listener but unless we get another
peer availability event we can't connect to the peer anymore because we
don't have a listener. So now we recreate the listener on recoverable
failures.

testCreatePeerListener - Tested that closing connections cleans things up
as expected and properly fires new listener events and cleaned up our
forward connection code to actually test things correctly.

testThaliMobileNativeWrapper - Test that we recreate listeners and they
can be used to reconnect at a higher level in the stack

createPeerListener - Fixed bugs in our logic for closing connections