WIP
testLocalSeqManager - Added a bunch of tests, the last one in the file is
failing because I don't properly return a promise in the case where
we are waiting for a timer to expire in order to run the next update. Will
fix tomorrow.

httpTester - Added utility to do GETs to a PSK server

localSeqManager - Fixing lots of bugs and adding functionality

utilities - Linting