Got localSeqManager tests passing!
testLocalSeqManagerCoordinated - Scoping issues, timing issues, etc.

testUtils - See the notes but there is a bug in PouchDB
(https://github.com/pouchdb/pouchdb/issues/5231) that was causing us
problems that I found a nasty work around for.

thaliNotificationServer - No functional changed, just refactored the code
to make it easier for me to intercept the function's output when I need
to log it.