Custom Express
I've been having some issues so I wanted to go from known state to known
state. This build will reliably pass unit tests but not coordinated tests
because it only contains our custom version of Express-PouchDB and not
PouchDB.

installCustomPouchDB - Detects if the local NPM registry (which has to
be set up manually) has the custom versions of the code we are looking
for. If it does then it exits. If not then it will pull down the repros,
build them and publish them to the local NPM registry.

package.json - Using our custom version of express-pouchdb but the public
version of pouchdb-node.

setUpDesktop - We have to run our script before npm install so our packages
are ready.

setUpTests - Same logic as above.