Get tests running again
I had made some changes intended to be finished by Artem but they killed
our tests and that turned out not to be useful. So I've fixed things up.

testThaliMobileNativeWrapper - The method name for terminateConnection
has become _terminateConnection and when thaliMobileNativeWrapper calls
tcpServersManager on a _terminateConnection call the equivalent method
in tcpServersManager is terminateIncomingConnection so I needed some glue
logic to fix this.

createPeerListener, thaliTcpServersManager - Put back the pleaseConnect
argument which needs to be removed as it isn't used with our new design.

thaliMobileNativeWrapper - Updated call to TCPServersManager to use the
new terminateIncomingConnection name.