Return multiconnect error on Android
We originally specified that if multiconnect was called on Android then
we wouldn't return the platform error in the sync multiconnect call but
rather as an async callback. That's a lot of work to do. To simplify it
we allow the platform error to be returned on the sync multiconnect call
so android doesn't have to implement the async callback.