It works on my machine :)
before_compile.js - Removed the code to edit the android manifest to force
a min sdk. I checked the compiled apk image and the min sdk set by the
Cordova environment variables are now being honored.

validateBuildEnvironment.js - Made Android checks based on the
$ANDROID_HOME environment variable, check the package.json in Thali for
the right btconnectorlib2 and jxcoreCordova versions, check that both
Node and JXCore are using alternate npm servers (aka Sinopia), checked that
the names in versions and CommandsAndResult are the same (to catch any
typos) and upgraded the file to ES6