We can get peerAvailability === false in Android
For some reason I thought that Android never sends a nonTCPPeerAvailability
event wiht availability === false. It turns out we do. We have a heuristic
where if we don't discover a peer for a minute then we will announce they
have gone. For now we will just accept this heuristic at face value and
honor it.