Validator and cleaned up sh scripts
build.sh - Removed the non-functional ERROR_ABORT (our use of -e meant
that the ERROR_ABORT function would never actually be called) and replaced
with a call to trap that will actually run. Also fixed a call to check if
CIGIVEMEMYIP.sh exists (to tell if we are in CI) which depended on being
able to fail that our new stricter environment setting don't like.

setUpDesktop.sh - Same as above

setUpTests.sh - Same as above but also had to set some default variables
because our new environment variables don't like unbound variables even
if they are positional.

validateBuildEnvironment.js - Updated git because old version doesn't work
well with worktrees. Put in a check to see if the configured NPM registry
actually will return a meaningful value. I had to do this because I can't
find a way to force a Sinopia server to identify itself as a Sinopia server.