fix document name generating
Sometimes first document is put in DB so fast that next document is
created with the same name because the name is just a current time in
milliseconds. In such a situation PouchDB raises conflict error. This
fix uses random prefix for test documents to avoid this kind of errors.