We aren't returning a promise!
When we close all the servers we accidentally called closeAll which is
a callback API as opposed to closeAllPromise which is a promise based
API. Which means the server would potentially still be open after we
thought it was closed. But I doubt this made much of a difference in
practice since we would have called close on the server thus stopping
it from accepting new connections and we would have called destroy
on all the connections both here and in the cleanups of the sockets and
connections in this file. But still.