Make coordination server run again
The actual bug was that I was in the middle of getting rid of desktop
as part of teaching the coordination server how to run both android and
iOS runs on the desktop and it screwed up Alex and Dersim who were using
this branch to test node against their native implementation.

HttpServer, Socket, TestDevice, TestFramework - I was getting an assert from my screwey config and couldn't
find out where because the asserts weren't giving me a stack so I had
to add a message to find the right one.

asserts - Taught it how to output an optional message for isObject

CoordinatedClient - Linting

runCoordinatedTests - This was the change that broke the coordination
server

UnitTest - This didn't actually break the test since this is used with
devices but I changed it back anyway

thaliMobile - Just changing this back to its original form, I had been
messing around with this code to find a problem in a previous commit