Clean up thaliMobile & make sure we properly ignore recreated events
testThaliMobile - Added a test to demonstrate that we properly clean up
thaliMobile after calling stop. Added a test to make sure that if a
peer is not available that recreated events will be ignored.

createPeerListener - Clearly marking when a failed connection is due to a
recreated action

thaliMobile - Making sure we clean up the cache after we stop and properly
pass through the peer.recreated event and use it to only update the cache
when thaliMobile has decided the peer is available.

thaliMobileNativeWrapper - Properly passing through the recreated property
and improving logging