Adding replication retry to wifi
build.sh - Alas there are some tests that are run in some configs and not
others so we have to run all three versions. But we really need to clean
up our tests to have 'common' and then 'wifi', 'native' and 'both' where
we only run tests out of 'common' that actually need a different network
layer.

testThaliNotification, testThaliNotificationClient,
testthaliPeerPoolOneAtATime, thaliNotificationClient - Added peerID to
PeerAdvertisesDataForUs

testUtils - Removed a secureOptions that somehow managed to survive the
purge

thaliTcpServersManager - Adding commentary

thaliReplicationPeerAction, thaliWifiInfrastructure - Linting

thaliPeerPoolOneAtATime - Added support for peerID being part of the
replication action and added support for retrying replications in Wifi.
Also fixed the retry behavior to not try to close a native session when
in Wifi.