#1588 Fixing up Android buffer problem and failing tests
SocketThreadBase - Turns out that the logic that got checked in could end up setting buffers to zero
because the Bluetooth socket get size calls could return 0. For now I decided to just remove the dynamic
size setting to simplify things. This should hopefully fix Android.

testThaliMobileNative - We have a condition where we could test something after we have called t.end but
this is necessary because our peer might be done with the test but the other might not. So now the test
will only cause a failure if it actually fails in which case we should fail.

testThaliMobileNativeWrapper - In addition to an identical issue to the one described above we also had tests
that were written before the skip functionality so I fixed that. Also some of the test names were so similar it
made doing searches hard so I added some text to make them different.

CoordinatedClient - Added the original result object to the error

makeIntoCloseAllServer - Linting