#1632 - Fixing broken build
README - Updating to the actual validateBuildEnvironment script that we now run

install - Turns out I used the syntax for build-extras.gradle instead of gradle.properties and didn't realize it because I still had the environment variables defined.

setUpDesktop - The logic we were using which would use the global variable if already defined otherwise it would set it doesn't work in the case of NVM_NODEJS_ORG_MIRROR if someone is using NVM which also sets this variable. Since NVM is more common than needing to change this location when doing our build I decided to revert that change and just overwrite the variable when we are doing a build.