#1620 - WIP - Adding native spec and code support for Android multicast lock
There is a bug in this code that causes a test to fail in testThaliMobile because the wifiBasedNativeMock changes its version of thaliWifiInfrastructure via proxyquire to just be a skeleton that is missing the methods I added here. So I have to fix that.

testthaliWifiInfrastructure - The check for platform.isMobile was wrong since that is always true. We needed to check for real mobile only. I also added two tests that make sure that we do call multicast lock on Android and don't call multicast lock anywhere else.

wifiBasedNativeMock - Added two thunk calls to fake calling the Android lock methods. And linting.

promiseQueue - Linting

thaliMobileNative - Defined how the native implementation of the Android multicast lock methods are to work. Also removed the definition of toggleWifi. This is actually defined by JXcore, not us and anyway per #1601 we are supposed to be out of this business and leave it to JXcore. Also linting.

thaliMobileNativeWrapper - Adding support for calling the multicast lock native call on Android

thaliWifiInfrastructure - Added the logic to get and release multicast locks on the Android platform. And a lot of linting.