Fix connection kill test
'#startUpdateAdvertisingAndListening - destroying the local connection
kills the connection to the remote peer' test destroys socket, then
tries to to connect to the same port again and expects this connection
to fail. But on desktop with mocked mobile layer reconnecting happens
too fast before everything is closed properly. So connection after close
has to be done asynchronously after the 'close' event