Clean up ThaliWifiInfrastructure
Original implementation of ThaliWifiInfrastructure consists of the
single entity that is responsible for several unrelated and independent
tasks: listening, advertising, network changes handling, async calls
management, and more.

The purpose of this change is to split these responsibilities between
different units, make it easier to understand and test, while providing
the original API to the upper layers.

So, the result is 3 independent entities:

- WifiAdvertiser - hosts router server and advertises current peer in
  the WiFi network
- WifiListener - listens for advertisements and filters
  invalid/irrelevant messages
- ThaliWifiInfrastructure - assembles advertiser and listener, queues
  async calls, and handles network changes